product filter refactor
